Add symlink check to makesymlinks.sh script
Without the check it overwrites the old dotfiles backups with symlinks
if the script is ran more than once. It also loops directories on MacOS

+++ makesymlinks.sh
@@ -7,7 +7,8 @@
 
 dir=~/dotfiles                        # dotfiles directory
 olddir=~/dotfiles_old                 # old dotfiles backup directory
-files="bash_profile bashrc tmux.conf liquidpromptrc" # list of files/folders to symlink in homedir
+files="bash_profile bashrc tmux.conf bash_it_custom" # list of files/folders to symlink in homedir
+
 
 ##########
 
@@ -25,8 +26,10 @@ echo "done"
 # then create symlinks from the homedir to any files in the ~/dotfiles
 # directory specified in $files
 for file in $files; do
-    echo "Moving any existing dotfiles from ~ to $olddir"
-    mv ~/.$file ~/dotfiles_old/
+    if [ ! -h ~/.$file ]; then # Move only if file is not already a symlink
+      echo "Moving any existing dotfiles from ~ to $olddir"
+      mv ~/.$file ~/dotfiles_old/
     echo "Creating symlink to $file in home directory."
     ln -s $dir/$file ~/.$file
+    fi
 done
